refactor(analyzer): split call.rs into focused submodules
call.rs at 1 504 lines was a single giant impl with three call patterns
and all type-compatibility helpers mixed together. Split into:

  call/mod.rs        — CallAnalyzer struct and public re-exports
  call/function.rs   — analyze_function_call
  call/method.rs     — analyze_method_call; extracts resolve_method_return
                       helper that eliminates the near-duplicate TNamedObject
                       / TSelf|TStaticObject|TParent branches (~270 lines)
  call/static_call.rs — analyze_static_method_call, analyze_static_dyn_method_call
  call/args.rs       — check_args, CheckArgsParams, check_constructor_args,
                       spread_element_type, substitute_static_in_return,
                       check_method_visibility, and all subtype helpers

No public API changes; all existing tests pass.
