+++ crates/mir-analyzer/src/call/mod.rs
@@ -1,0 +1,8 @@
+mod args;
+mod function;
+mod method;
+mod static_call;
+
+pub use args::{check_constructor_args, spread_element_type, CheckArgsParams};
+
+pub struct CallAnalyzer;
